[TECH][HTCHT-1732] ensure generated deployment names and vm names are unique
Azure deployment names are generated using a millisecond-precision
timestamp to guarantee uniqueness. This can fail in cases where multiple
jobs have an identical trigger, resulting in conflicting deployment
names and resource names, which in turns results in errors in a variety
of systems expecting these identifiers to be unique.

We remedy this by overwriting the milliseconds part of the timestamp
with a thread-safe per-second counter, preserving the timestamp nature
of the identifier but guaranteeing that we don't use duplicate
timestamps.